Split common timeline functions out
